CardanoCLI: add policy creation

+++ apps/Csharp.CardanoSounds/CS.Csharp.CardanoCLI/Models/TransactionParams.cs
@@ -11,11 +11,11 @@ namespace CS.Csharp.CardanoCLI.Models
 
         public string TxInHash { get; set; }
 
-        public int TxInLovelaceValue { get; set; }
+        public long TxInLovelaceValue { get; set; }
 
         public int TxInIx { get; set; }
 
-        public int LovelaceValue { get; set; }
+        public long LovelaceValue { get; set; }
 
         public string SenderAddress { get; set; }
